Add files via upload
Welcome to go go fish. For this game, there are two players, one is a human interacting and
another is the computer. Each player starts with 7 cards and the rest of the cards are kept on
the deck. The goal of the game is to make matches of 4 cards. Each group of four cards is
called a book. The player with the most books wins the game.
In the beginning of the turn, the player goes first and names a card and asks the computer if it
has the card. The player does so in order to try to make a match of 4 of that card. If the
computer has a card, it has to give it to the player. The player asks for another card from the
computer to try to make another match. If the computer doesn’t have the card player asks, it
says ‘go fish’ and the player has to draw one card from the deck and now it’s the computer’s
turn. Same thing happens until one player has 7 books or the deck is empty, then the player with
the most number of books wins.